+++ src/game/player.ts
@@ -20,6 +20,16 @@ export class Player extends Sprite {
     statuses;
     nextLevel: number | undefined; // where to go next frame
     realPlayer: boolean; // whether or not player is the actual one. Controls changing lvls and stuff
+    gravityMult: number;
+
+    // which way gravity is. Positive means falling down, negative falls up
+    get gravityDir() {
+        return Math.sign(this.gravityMult);
+    }
+
+    set gravityDir(val: number) {
+        this.gravityMult = Math.abs(this.gravityMult) * val;
+    }
 
     constructor(x: number, y: number, real = true) {
         super(Assets.get('player'), x, y);
@@ -28,6 +38,7 @@ export class Player extends Sprite {
         this.wallJumpCooldown = 0;
         this.statuses = new Set<string>();
         this.realPlayer = real;
+        this.gravityMult = 1;
 
         this.vy = -3; // parity with original
     }
@@ -43,9 +54,9 @@ export class Player extends Sprite {
         );
     }
 
-    colliding() {
+    checkCollision(firstCheck?: boolean) {
         for (const block of World.data) {
-            if (block.colliding(this)) return true;
+            if (block.colliding(this, firstCheck)) return true;
         }
         return false;
     }
@@ -61,24 +72,31 @@ export class Player extends Sprite {
     physicsTick() {
         const movementVector = this.getMovementVector();
 
-        this.vy += PlayerSettings.gravity;
+        this.vy += PlayerSettings.gravity * this.gravityMult;
 
         // slow gravity if too fast
-        if (this.vy > PlayerSettings.maxFallSpeed) {
-            this.vy--;
+        if (this.vy * this.gravityDir > PlayerSettings.maxFallSpeed) {
+            this.vy -= this.gravityMult;
         }
 
-        // TODO: Bounce block stuff happens here i think
-        this.colliding();
+        // this is to trigger effects such as bounce blocks
+        this.checkCollision(true);
 
         this.y += this.vy;
 
-        // Some mud check
+        // Some mud check needs to go here
+        this.checkCollision(); 
+        // TODO: try removing this checkCollision and see if the mud still behaves the same
+        //       I am not sure if checking collision again after y movement will really change that much
+
+        if (this.hasStatus('mud')) {
+            this.vy = 0;
+        }
 
         let previouslyTouchedIce = this.hasStatus('icy');
 
         // slowdown if not moving
-        if (movementVector.x === 0 && !this.hasStatus('icy')) {
+        if ((movementVector.x === 0 && !this.hasStatus('icy')) || this.hasStatus('mud')) {
             this.vx -= Math.sign(this.vx);
         }
         if (Math.abs(this.vx) > PlayerSettings.speedHardcap) {
@@ -93,24 +111,24 @@ export class Player extends Sprite {
         }
 
         // some goofty ahh collision code
-        if (this.colliding()) {
-            if (this.vy > 0) { // if falling
-                // TODO: whatever rjump is it goes here
-                this.canJump = true;
+        if (this.checkCollision()) {
+            if (this.vy * this.gravityMult > 0 || this.hasStatus('mud')) { // if falling
+                if (!this.hasStatus('no-jump-regen'))
+                    this.canJump = true;
 
                 this.statuses.delete('icy');
                 previouslyTouchedIce = false;
 
                 // Go backwards until no longer in floor
                 for (let i = 0; i < PlayerSettings.maxFloorHeight; i++) {
-                    if (!this.colliding()) break;
-                    this.y -= 1;
+                    if (!this.checkCollision()) break;
+                    this.y -= this.gravityDir; // TODO: make a var outside the loop instead of calling Math.sign repeatedly?
 
                     // Stuck in wall, so lets just go down
                     if (i === PlayerSettings.maxFloorHeight - 1) {
-                        this.y += PlayerSettings.maxFloorHeight;
-                        while (this.colliding()) {
-                            this.y++;
+                        this.y += PlayerSettings.maxFloorHeight * this.gravityDir;
+                        while (this.checkCollision()) {
+                            this.y += this.gravityDir;
                         }
                     }
                 }
@@ -119,24 +137,24 @@ export class Player extends Sprite {
 
             } else { // if not falling (either swapped while in wall or is hitting ceiling)
                 for (let i = 0; i < PlayerSettings.maxHorizontalNudge; i++) { // lets try nudging u to the left
-                    if (!this.colliding()) break;
+                    if (!this.checkCollision()) break;
                     this.x--;
                     if (i === PlayerSettings.maxHorizontalNudge - 1) this.x += PlayerSettings.maxHorizontalNudge;
                 }
 
                 for (let i = 0; i < PlayerSettings.maxHorizontalNudge; i++) { // lets try nudging u to the right
-                    if (!this.colliding()) break;
+                    if (!this.checkCollision()) break;
                     this.x++;
                     if (i === PlayerSettings.maxHorizontalNudge - 1) this.x -= PlayerSettings.maxHorizontalNudge;
                 }
 
-                while (this.colliding()) this.y++; // nudging is not working activate nuclear mode and send u down
+                while (this.checkCollision()) this.y += this.gravityDir; // nudging is not working activate nuclear mode and send u down
             }
             this.vy = 0;
         }
 
         // If falling, then you are not on the ground which means u can't jump!
-        if (this.vy > PlayerSettings.coyoteFrames) {
+        if (this.vy * this.gravityDir > PlayerSettings.coyoteFrames) { // TODO: velocity maybe should be divided by this.gravityMult for this check. Coyote frames should be consistent across different gravity strengths
             this.canJump = false;
         }
 
@@ -147,40 +165,30 @@ export class Player extends Sprite {
             this.wallJumpCooldown--;
         }
 
+        let speed = PlayerSettings.speed;
+        if (this.hasStatus('icy')) speed += 3;
+        if (this.hasStatus('mud')) speed -= 3;
+
         // limit x velocity
-        if (Math.abs(this.vx) > PlayerSettings.speed + (this.hasStatus('icy') ? 3 : 0) && this.wallJumpCooldown === 0) { // TODO: allow for temp changing player speed so I don't put this hasStatus ternary in the conditional
+        if (Math.abs(this.vx) > speed && this.wallJumpCooldown === 0) { // TODO: allow for temp changing player speed so I don't put this hasStatus ternary in the conditional
             this.vx -= Math.sign(this.vx);
         }
 
-        // change x
-        this.x += this.vx;
-        // add collision checking
-        for (let i = 0; i < PlayerSettings.maxSlopeHeight; i++) {
-            if (!this.colliding()) break;
-
-            this.y--;
-            if (i === PlayerSettings.maxSlopeHeight - 1) {
-                this.y += PlayerSettings.maxSlopeHeight;
-                while(this.colliding()) this.x -= Math.sign(this.vx);
-
-                if (movementVector.y === -1 && Math.abs(this.vx) > 2) {
-                    if (this.wallJumpCooldown < 3 && !this.hasStatus('no-wj')) { // some njump thing goes here
-                        this.vx = Math.round(this.vx * -1.2);
-                        this.vy = Math.round(PlayerSettings.jumpStrength / -1.5);
-                        this.wallJumpCooldown = 7; // 7 frame cooldown
-                    }
-                } else {
-                    this.vx = 0;
-                }
-            }
-        }
+        this.updateXPos(movementVector);
 
         // particle stuff here
 
         // jump
-        if (movementVector.y === -1 && this.canJump) {
-            this.vy = -PlayerSettings.jumpStrength;
-            this.canJump = false;
+        if (movementVector.y === -1) {
+            if (this.hasStatus('mud'))
+            {
+                if (Input.justGet('up'))
+                    this.jump();
+            } 
+            else if (this.canJump)
+            {
+                this.jump();
+            }
         }
 
         // New level stuff
@@ -205,5 +213,35 @@ export class Player extends Sprite {
 
         if (previouslyTouchedIce)
             this.statuses.add('icy');
+    }
+
+    updateXPos(movementVector: Vector2) {
+        // change x
+        this.x += this.vx;
+        // add collision checking
+        for (let i = 0; i < PlayerSettings.maxSlopeHeight; i++) {
+            if (!this.checkCollision()) break;
+
+            this.y -= this.gravityDir;
+            if (i === PlayerSettings.maxSlopeHeight - 1) {
+                this.y += PlayerSettings.maxSlopeHeight * this.gravityDir;
+                while(this.checkCollision()) this.x -= Math.sign(this.vx);
+
+                if (movementVector.y === -1 && Math.abs(this.vx) > 2) {
+                    if (this.wallJumpCooldown < 3 && !this.hasStatus('no-wj')) { // some njump thing goes here
+                        this.vx = Math.round(this.vx * -1.2);
+                        this.vy = Math.round(PlayerSettings.jumpStrength / -1.5 * this.gravityDir);
+                        this.wallJumpCooldown = 7; // 7 frame cooldown
+                    }
+                } else {
+                    this.vx = 0;
+                }
+            }
+        }
+    }
+
+    jump() {
+        this.vy = -PlayerSettings.jumpStrength * this.gravityDir;
+        this.canJump = false;
     }
 }