+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.8.0 January 9, 2026
+- Added
+    - Trials World
+    - Interlude world
+    - Final 4 levels in world 3
+    - Race of keys
+    - Vine block
+    - Bounce left, right, and down blocks
+    - Mud block
+    - Nojump block
+    - Small block
+    - Death block
+    - Reverse and normalize gravity blocks
+    
+- Fixed
+    - Bounce blocks no longer push during collision checks. This is a problem when halfway on a bounce and halfway on a solid block,
+    during the collision check the player would temporarily be lowered into the bounce block, triggering it.
+    - Format 2 levels with doors/keys no longer are broken
+
 ## 0.7.1 January 8, 2026
 - Fixed levels in world 3 not having a format, which stopped the level from being loaded
     - Tower of Ratio